feat(hdf-10): leaderboard table with filters work in progress
