+++ packages/apps/dashboard/ui-2024/src/components/Home/index.ts
@@ -1,2 +1,3 @@
 export * from './SmallGraph';
 export * from './GraphSwiper';
+export * from './Leaderboard';
