+++ packages/apps/dashboard/ui-2024/src/pages/Home/Home.tsx
@@ -1,4 +1,3 @@
-import React, { useState } from 'react';
 import PageWrapper from '@components/PageWrapper';
 import Search from '@components/Search';
 import ShadowIcon from '@components/ShadowIcon';
@@ -7,22 +6,10 @@ import Typography from '@mui/material/Typography';
 import Tooltip from '@mui/material/Tooltip';
 import Button from '@mui/material/Button';
 import HelpOutlineIcon from '@mui/icons-material/HelpOutline';
-import Table from '@mui/material/Table';
-import TableBody from '@mui/material/TableBody';
-import TableCell, { tableCellClasses } from '@mui/material/TableCell';
-import TableContainer from '@mui/material/TableContainer';
-import TableHead from '@mui/material/TableHead';
-import TableRow from '@mui/material/TableRow';
-import Paper from '@mui/material/Paper';
-import Grid from '@mui/material/Grid';
-import InputLabel from '@mui/material/InputLabel';
-import MenuItem from '@mui/material/MenuItem';
-import FormControl from '@mui/material/FormControl';
-import Select, { SelectChangeEvent } from '@mui/material/Select';
 import Divider from '@mui/material/Divider';
 
 import SimpleBar from 'simplebar-react';
-import { Link, useNavigate } from 'react-router-dom';
+import { Link } from 'react-router-dom';
 
 import bitfinex from '@assets/bitfinex.png';
 import probitGlobal from '@assets/probitGlobal.png';
@@ -31,301 +18,10 @@ import bing from '@assets/bing.png';
 import coinlist from '@assets/coinlist.png';
 import lbank from '@assets/lbank.png';
 import cup from '@assets/cup.png';
-import human from '@assets/human.png';
-import exchange from '@assets/exchange.png';
-import recording from '@assets/recording.png';
-import reputation from '@assets/reputation.png';
-import AbbreviateClipboard from '@components/SearchResults/AbbreviateClipboard';
-import MoonbaseAlphaIcon from '@components/Icons/MoonbaseAlphaIcon';
-
-import MoonbeamIcon from '@components/Icons/MoonbeamIcon';
-import BinanceSmartChainIcon from '@components/Icons/BinanceSmartChainIcon';
-import EthereumIcon from '@components/Icons/EthereumIcon';
-import HumanIcon from '@components/Icons/HumanIcon';
-import PolygonIcon from '@components/Icons/PolygonIcon';
-import CeloIcon from '@assets/icons/celo.svg';
-import SvgIcon from '@mui/material/SvgIcon';
+import Leaderboard from '@components/Home/Leaderboard';
 import GraphSwiper from '@components/Home/GraphSwiper';
 
-function createData(
-	role: string,
-	address: string,
-	stake: string,
-	networks: string,
-	reputation: string,
-	operator: string
-) {
-	return { role, address, stake, networks, reputation, operator };
-}
-
-const rows = [
-	createData(
-		'Job Launcher',
-		'0x67499f129433b82e5a4e412143a395e032e76c0dc0f83606031',
-		'1e-18 HMT',
-		'Polygon',
-		'Medium',
-		'1%'
-	),
-	createData(
-		'Recording Oracle',
-		'0x67499f129433b82e5a4e412143a395e032e76c0dc0f83606031',
-		'1e-18 HMT',
-		'Polygon',
-		'High',
-		'1%'
-	),
-	createData(
-		'Reputation Oracle',
-		'0x67499f129433b82e5a4e412143a395e032e76c0dc0f83606031',
-		'3e-18 HMT',
-		'Polygon',
-		'Medium',
-		'2%'
-	),
-	createData(
-		'Exchange Oracle',
-		'0x67499f129433b82e5a4e412143a395e032e76c0dc0f83606031',
-		'4e-18 HMT',
-		'Polygon',
-		'Low',
-		'1%'
-	),
-	createData(
-		'HUMAN App',
-		'0x67499f129433b82e5a4e412143a395e032e76c0dc0f83606031',
-		'2e-18 HMT',
-		'Polygon',
-		'Coming soon',
-		'5%'
-	),
-];
-
-interface Item {
-	role: string;
-	address: string;
-	stake: string;
-	networks: string;
-	reputation: string;
-	operator: string;
-}
-
-type networkTypes =
-	| 'ethereum'
-	| 'goerli'
-	| 'binance'
-	| 'testnet'
-	| 'polygon'
-	| 'mumbai'
-	| 'moonbeam'
-	| 'all'
-	| 'alpha';
-
 const Home: React.FC = () => {
-	const [network, setNetwork] = useState<networkTypes>('all');
-	const navigate = useNavigate();
-
-	const handleChange = (event: SelectChangeEvent) => {
-		setNetwork(event.target.value as networkTypes);
-	};
-
-	const renderIcon: React.FC<Item> = (item) => {
-		let src = '';
-		switch (item.role) {
-			case 'Job Launcher':
-				return (
-					<div className="icon-table">
-						<span>JL</span>
-					</div>
-				);
-			case 'Recording Oracle':
-				src = recording;
-				break;
-			case 'Reputation Oracle':
-				src = reputation;
-				break;
-			case 'Exchange Oracle':
-				src = exchange;
-				break;
-			case 'HUMAN App':
-				src = human;
-				break;
-			default:
-				src = human;
-				break;
-		}
-
-		return (
-			<div className="icon-table">
-				<img src={src} alt="logo" />
-			</div>
-		);
-	};
-
-	//TODO rendering all images and text when all networkType is selected
-	const renderNetworkDetails = (network: networkTypes) => {
-		const networkDetails = {
-			ethereum: {
-				title: 'Ethereum',
-				icon: <EthereumIcon />,
-			},
-			goerli: {
-				title: 'Ethereum Goreli',
-				icon: <EthereumIcon />,
-			},
-			binance: {
-				title: 'Binance Smart Chain',
-				icon: <BinanceSmartChainIcon />,
-			},
-			testnet: {
-				title: 'Smart Chain (Testnet)',
-				icon: <BinanceSmartChainIcon />,
-			},
-			polygon: {
-				title: 'Polygon',
-				icon: <PolygonIcon />,
-			},
-			mumbai: {
-				title: 'Polygon Mumbai',
-				icon: <PolygonIcon />,
-			},
-			moonbeam: {
-				title: 'Moonbeam',
-				icon: <MoonbeamIcon />,
-			},
-			alpha: {
-				title: 'Moonbase Alpha',
-				icon: <MoonbeamIcon />,
-			},
-			celo: {
-				title: 'Celo',
-				icon: <CeloIcon />,
-			},
-			alfajores: {
-				title: 'alfajores',
-				icon: <CeloIcon />,
-			},
-		};
-
-		if (network === undefined) {
-			return null;
-		}
-
-		if (network === 'all') {
-			return (
-				<>
-					{networkDetails['alpha'].icon}
-					{networkDetails['ethereum'].title}
-				</>
-			);
-		}
-
-		return (
-			<>
-				{networkDetails[network].icon}
-				{networkDetails[network].title}
-			</>
-		);
-	};
-
-	const renderReputation: React.FC<Item> = (item) => {
-		switch (item.reputation) {
-			case 'Medium':
-				return (
-					<div className="reputation-table reputation-table-medium">
-						{item.reputation}
-					</div>
-				);
-			case 'High':
-				return (
-					<div className="reputation-table reputation-table-high">
-						{item.reputation}
-					</div>
-				);
-			case 'Low':
-				return (
-					<div className="reputation-table reputation-table-low">
-						{item.reputation}
-					</div>
-				);
-			case 'Coming soon':
-				return (
-					<div className="reputation-table reputation-table-soon">
-						{item.reputation}
-					</div>
-				);
-			default:
-				return (
-					<div className="reputation-table reputation-table-soon">
-						{item.reputation}
-					</div>
-				);
-		}
-	};
-
-	const selectRender = () => {
-		return (
-			<FormControl fullWidth size="small">
-				<InputLabel id="network-select-label">By Network</InputLabel>
-				<Select
-					labelId="network-select-label"
-					id="network-select"
-					value={network}
-					label="By Network"
-					onChange={handleChange}
-				>
-					<MenuItem className="select-item" value={'all'}>
-						<HumanIcon />
-						All Networks
-					</MenuItem>
-					<MenuItem className="select-item" value={'ethereum'}>
-						<EthereumIcon />
-						Ethereum
-					</MenuItem>
-					<MenuItem className="select-item" value={'goerli'}>
-						<EthereumIcon />
-						Ethereum Goerli
-					</MenuItem>
-					<MenuItem className="select-item" value={'binance'}>
-						<BinanceSmartChainIcon />
-						Binance Smart Chain
-					</MenuItem>
-					<MenuItem className="select-item" value={'testnet'}>
-						<BinanceSmartChainIcon /> Smart Chain (Testnet)
-					</MenuItem>
-					<MenuItem className="select-item" value={'polygon'}>
-						<PolygonIcon />
-						Polygon
-					</MenuItem>
-					<MenuItem className="select-item" value={'mumbai'}>
-						<PolygonIcon />
-						Polygon Mumbai
-					</MenuItem>
-					<MenuItem className="select-item" value={'moonbeam'}>
-						<MoonbeamIcon />
-						Moonbeam
-					</MenuItem>
-					<MenuItem className="select-item" value={'alpha'}>
-						<MoonbaseAlphaIcon />
-						Moonbase Alpha
-					</MenuItem>
-					<MenuItem className="select-item" value={'celo'}>
-						<SvgIcon>
-							<CeloIcon />
-						</SvgIcon>
-						Celo
-					</MenuItem>
-					<MenuItem className="select-item" value={'alfajores'}>
-						<SvgIcon>
-							<CeloIcon />
-						</SvgIcon>
-						Celo Alfajores
-					</MenuItem>
-				</Select>
-			</FormControl>
-		);
-	};
-
 	return (
 		<PageWrapper violetHeader>
 			<div className="home-page-header">
@@ -457,98 +153,7 @@ const Home: React.FC = () => {
 				img={cup}
 			/>
 
-			<TableContainer
-				component={Paper}
-				sx={{ padding: '32px', marginTop: '30px' }}
-			>
-				<div className="mobile-select">{selectRender()}</div>
-				<SimpleBar>
-					<Table
-						sx={{
-							minWidth: 650,
-							[`& .${tableCellClasses.root}`]: {
-								borderBottom: 'none',
-							},
-						}}
-						aria-label="simple table"
-					>
-						<TableHead>
-							<TableRow className="home-page-table-header">
-								<TableCell>ROLE</TableCell>
-								<TableCell>
-									<div className="icon-table">
-										<Tooltip title="Address of the role" arrow>
-											<HelpOutlineIcon color="sky" />
-										</Tooltip>
-										<span>ADDRESS</span>
-									</div>
-								</TableCell>
-								<TableCell>
-									<div className="icon-table">
-										<Tooltip title="Amount of HMT staked" arrow>
-											<HelpOutlineIcon color="sky" />
-										</Tooltip>
-										<span>STAKE</span>
-									</div>
-								</TableCell>
-								<TableCell className="table-filter-select">
-									{selectRender()}
-									<span className="mobile-title">NETWORK</span>
-								</TableCell>
-								<TableCell>
-									<div className="icon-table">
-										<Tooltip
-											title="Reputation of the role as per their activities "
-											arrow
-										>
-											<HelpOutlineIcon color="sky" />
-										</Tooltip>
-										<span>REPUTATION SCORE</span>
-									</div>
-								</TableCell>
-								<TableCell>OPERATOR FEE</TableCell>
-							</TableRow>
-						</TableHead>
-						<TableBody>
-							{rows.map((row) => (
-								<TableRow
-									onClick={() => navigate(`/search/${row.address}`)}
-									key={row.role}
-									className="home-page-table-row"
-								>
-									<TableCell>
-										{renderIcon(row)}
-										{row.role}
-									</TableCell>
-									<TableCell>
-										<Grid
-											container
-											wrap="nowrap"
-											alignItems="center"
-											sx={{ gap: '18px' }}
-										>
-											<AbbreviateClipboard value={row.address} />
-										</Grid>
-									</TableCell>
-									<TableCell>{row.stake}</TableCell>
-									<TableCell>
-										<Grid
-											container
-											wrap="nowrap"
-											alignItems="center"
-											sx={{ gap: '18px' }}
-										>
-											{renderNetworkDetails(network)}
-										</Grid>
-									</TableCell>
-									<TableCell>{renderReputation(row)}</TableCell>
-									<TableCell>{row.operator}</TableCell>
-								</TableRow>
-							))}
-						</TableBody>
-					</Table>
-				</SimpleBar>
-			</TableContainer>
+			<Leaderboard />
 		</PageWrapper>
 	);
 };
